Moved Header out of dashboard module. - Added Primary website link on logo. - Renamed Dashboard component. - Renamed Header Component.
