+++ src/modules/uv_dashboard/uv_dashboard.actions.ts
@@ -1,5 +1,5 @@
 import UV_DASHBOARD from "./uv_dashboard.constants";
-import { UVDashboardType } from './../../shared/Types';
+import { UVDashboardType } from '../../shared/Types';
 
 const initDashboard = () => {
   return {
