+++ src/modules/uv_dashboard/uv_dashboard.reducer.ts
@@ -3,7 +3,7 @@ import UV_PIE from "../../components/uv_pie/uv_pie.constants";
 import { UVAction } from '../../shared/Types';
 import { UVItem } from '../../shared/Types';
 import UV_DASHBOARD from "./uv_dashboard.constants";
-import { UVCategory, UVNumberProps } from './../../shared/Types';
+import { UVCategory, UVNumberProps } from '../../shared/Types';
 import { mapNumberComponents } from "./uv_dashboard.saga";
 
 const initialState = {
