+++ src/root.reducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
-import UvDashboardReducer from './modules/dashboard/uv_dashboard.reducer';
 import uvHeaderReducer from './components/uv_header/uv_header.reducers';
+import UvDashboardReducer from './modules/uv_dashboard/uv_dashboard.reducer';
 
 export const rootReducer = combineReducers({
   header: uvHeaderReducer,
