+++ src/root.sagas.ts
@@ -1,6 +1,6 @@
 import { SagaMiddleware } from 'redux-saga';
 
-import { UvDashboardSaga } from './modules/dashboard/uv_dashboard.saga';
+import { UvDashboardSaga } from './modules/uv_dashboard/uv_dashboard.saga';
 import { uvHeaderSaga } from './components/uv_header/uv_header.saga';
