+++ src/uv_store.ts
@@ -7,7 +7,7 @@ import * as headerData from './components/uv_header/uv_header.json';
 import { rootReducer } from './root.reducer';
 import { runAllSaga } from './root.sagas';
 
-import { initDashboard } from './modules/dashboard/uv_dashboard.actions';
+import { initDashboard } from './modules/uv_dashboard/uv_dashboard.actions';
 import { loadHeader } from './components/uv_header/uv_header.actions';
 
 const sagaMiddleware = createSagaMiddleware();
@@ -19,5 +19,5 @@ export const uvStore = createStore(
 
 runAllSaga(sagaMiddleware);
 
-uvStore.dispatch(initDashboard());
 uvStore.dispatch(loadHeader(headerData.config));
+uvStore.dispatch(initDashboard());
